Support different discount levels for prices
Convert database so that each research level can have store different
price for each discount level.

Modify application to use new database format and support calculating
estimated price from if exact price for current discount level is not
known.

If shown cost is estimated '*' is shown next to the price to indicate
that.