Refactor levels
Change levels in database to be numeric level instead of the previous
percentage/multiplier value.

This simplifies code a lot.

With this also next_level column is removed from UI as it's useless
information now.

Also fix the estimate calculation for missing prices in database which
became lot easier with this level change.